feat: 사용량 과금을 계산하는 로직 작성

+++ src/main/java/org/example/moono_backend/domain/Plan.java
@@ -26,7 +26,7 @@ public class Plan extends BaseEntity {
     private Integer baseFee;
 
     //기본 제공량
-    private Double basicMobileData;
+    private Integer basicMobileData;
     private Integer basicVoice;
     private Integer basicSms;
